+++ VersionInfo.cs
@@ -3,11 +3,11 @@ using System.Reflection;
 [assembly: AssemblyTitle("FreakShow")]
 [assembly: AssemblyProduct("FreakShow")]
 [assembly: AssemblyCompany("kappiderfreak")]
-[assembly: AssemblyVersion("1.2.4.0")]
-[assembly: AssemblyFileVersion("1.2.4.0")]
-[assembly: AssemblyInformationalVersion("1.2.4")]
+[assembly: AssemblyVersion("1.2.5.0")]
+[assembly: AssemblyFileVersion("1.2.5.0")]
+[assembly: AssemblyInformationalVersion("1.2.5")]
 
 internal static class FreakShowVersion
 {
-    public const string Current = "1.2.4";
+    public const string Current = "1.2.5";
 }